+++ app/examples/react-17/src/index.js
@@ -1,10 +1,7 @@
+import 'dashboard-example/index.css'
 import React from 'react'
 import ReactDOM from 'react-dom'
-
 import App from './App'
-
-import 'styles/global.css'
-import 'styles/index.css'
 
 const root = document.getElementById('root')
 ReactDOM.render(<App />, root)