+++ app/examples/react-18/src/index.js
@@ -1,10 +1,7 @@
-import React from 'react'
+import 'dashboard-example/index.css'
 import { createRoot } from 'react-dom/client'
-
+import React from 'react'
 import App from './App'
-
-import 'styles/global.css'
-import 'styles/index.css'
 
 const root = createRoot(document.getElementById('root'))
 root.render(<App />)