+++ packages/ui-kit/src/helpers/chartUtils/chartUtils.ts
@@ -132,19 +132,19 @@ export const getCustomChartLabelsPlugin = ({
 
       if (datasetMeta.type === 'doughnut' && !hideTotal) {
         const totalValue = dataset.data.reduce((a: number, c: number) => a + c, 0).toLocaleString()
+
+        ctx.save()
+
+        const xCoor = chart.getDatasetMeta(0).data[0]?.x
+        const yCoor = chart.getDatasetMeta(0).data[0]?.y
+
+        ctx.textAlign = 'center'
+        ctx.textBaseline = 'middle'
         ctx.font = `12px ${theme?.tinyFontFamily}`
-        ctx.fillText(
-          'Total',
-          chart.width / 2 - ctx.measureText('Total').width / 2,
-          chart.chartArea.height / 2 + chart.chartArea.top - 20
-        )
+        ctx.fillText('Total', xCoor, yCoor - 12)
 
         ctx.font = `700 24px ${theme?.tinyFontFamily}`
-        ctx.fillText(
-          totalValue,
-          chart.width / 2 - ctx.measureText(totalValue).width / 2,
-          chart.chartArea.height / 2 + chart.chartArea.top + 5
-        )
+        ctx.fillText(totalValue, xCoor, yCoor + 12)
       }
     }
   }